string

+++ main.h
@@ -7,7 +7,7 @@ int _print_strin(char *c, int *s, int *i);
 int _putchar(char c);
 int count_size(long int c);
 int _bool(const char *format, char *spec, int i);
-int _print_number2(long int c);
+int _print_number2(int c);
 
 #include <unistd.h>
 #include <stdio.h>
